requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
httpx
python-dotenv
python-multipart
pydantic-ai[gemini]

